Добавил шаблонизацию
